Adding usePreviewHandler and useRenderHandler hook
